requirements.txt
transformers>=4.44.0
peft>=0.12.0
torch>=2.6.0
datasets>=2.21.0
accelerate>=1.0.0
bitsandbytes>=0.45.0
boto3>=1.35.0
sagemaker>=2.232.0,<3
fastapi>=0.119.0
uvicorn>=0.30.5
streamlit>=1.38.0
plotly>=5.23.0
python-dotenv>=1.0.1
requests>=2.32.3
tqdm>=4.66.5
filelock>=3.15.4
